When box does not has "/proc/stb/vmpeg/1/visible" do not stop PIP
Previously when a box did not have the pip visible proc and for a PIG
screen the PIP needs to be hidden the PIP service was stopped.
This means when you open the e.g. the channelselection screen in a PIG skin
the PIP was stopped and this means tuners that are actually not available
became available. So when you leave the channelselection screen you can
get a tuner not available message as soon the PIP and a new selected
service should be shown.

Now instead of stopping the service hide the PIP in a different way by making
it small and relocating it to the left top of the screen so it will also be
invisible but in a different way. I need to make it 2x2 pixes (some hardware
do crash when I make it 0,0 in size) but you still do not see it. When the
PIG screen is left the PIP will be restored to the orriginal size and
location